+++ CotEditor/Sources/FindPanelTextView.swift
@@ -172,3 +172,38 @@ final class FindPanelTextView: NSTextView {
     }
     
 }
+
+
+
+extension FindPanelTextView {
+    
+    /// adjust word selection range
+    override func selectionRange(forProposedRange proposedCharRange: NSRange, granularity: NSSelectionGranularity) -> NSRange {
+        
+        let range = super.selectionRange(forProposedRange: proposedCharRange, granularity: granularity)
+        
+        guard UserDefaults.standard[.findUsesRegularExpression] else { return range }
+        
+        guard granularity == .selectByWord else { return range }
+        
+        // settle result on expanding selection or if there is no possibility for clicking a bracket
+        guard proposedCharRange.length == 0, range.length == 1 else { return range }
+        
+        let characterIndex = Range(range, in: self.string)!.lowerBound
+        
+        // select inside of brackets
+        if let pairIndex = self.string.indexOfBracePair(at: characterIndex, candidates: [BracePair("(", ")"), BracePair("[", "]")], ignoring: BracePair("[", "]")) {
+            switch pairIndex {
+            case .begin(let beginIndex):
+                return NSRange(beginIndex...characterIndex, in: self.string)
+            case .end(let endIndex):
+                return NSRange(characterIndex...endIndex, in: self.string)
+            case .odd:
+                return NSRange(characterIndex...characterIndex, in: self.string)
+            }
+        }
+        
+        return range
+    }
+    
+}